requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Deprecated==1.2.14
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gcloud==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
h5py==3.9.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
jws==0.1.3
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyparsing==3.0.9
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-jwt==2.0.1
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
whitenoise==6.5.0
wrapt==1.15.0
